+++ tests/test_catalog_maker/test_catalog_maker.py
@@ -188,10 +188,8 @@ class TestCatalogMaker:
 
         assert rh.count_failures() == 0
 
-    @pytest.mark.skipif(
-        os.environ.get("ABCUNIT_DB_SETTINGS") is None, reason="database backend not set"
-    )
     @classmethod
     def teardown_class(cls):
-        rh = DataBaseHandler(table_name="c3s_cmip6_test_catalog_results")
-        rh._delete_table()
+        if os.environ.get("ABCUNIT_DB_SETTINGS"):
+            rh = DataBaseHandler(table_name="c3s_cmip6_test_catalog_results")
+            rh._delete_table()