+++ HISTORY.rst
@@ -5,9 +5,10 @@ v0.4.0 (unreleased)
 -------------------
 Breaking Changes
 ^^^^^^^^^^^^^^^^
-* ``pandas`` now a dependency.
 * Inventory maker now removed and replaced by intake catalog maker which writes a csv file with the dataset entries and a yaml description file.
 * In ``etc/roocs.ini`` the option ``use_inventory`` has been replaced by ``use_catalog`` and the inventory maker options have been replaced with equivalent catalog options. However, the option to include file paths or not no longer exists.
+* The catalog maker now uses a database backend and creates a csv file so there are 2 new dependencies for the catalog maker: pandas and abcunit-backend.
+This means a database backend must be specified and the paths for the pickle files in ``etc/roocs.ini`` are no longer necessary. For more information see the README.
 
 Other Changes
 ^^^^^^^^^^^^^